allow programmatically calling codegen
