+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@backengine/codegen",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "description": "Generate code for Backengine projects.",
   "bin": "build/index.js",
   "files": [
